cleanup: move FullScreenModal component to atoms folder
